[UPLOAD]: add adversarial examples
Adversarial samples are created by our attack for each perturbation instruction and each enhancing strategy. _POS means that we selected only the initial positive sentences and tried to modify them to have a negative sentiment. _NEG means that we selected only the initial negative sentences and tried to modify them to have a positive sentiment